디자인 QA 반영
