+++ app/about/future-careers/page.tsx
@@ -59,7 +59,11 @@ function CompanyTableRow({ index, name, url, year }: CompanyTableRowProps) {
     <li className="flex pt-[.63rem] pb-3 odd:bg-neutral-50">
       <p className={'pl-5 ' + TABLE_COLUMN_SIZE[0]}>{index}</p>
       <p className={'pl-3 ' + TABLE_COLUMN_SIZE[1]}>{name}</p>
-      <a className={'text-link pl-3 ' + TABLE_COLUMN_SIZE[2]} href={url} target="_blank">
+      <a
+        className={'text-link pl-3 hover:underline ' + TABLE_COLUMN_SIZE[2]}
+        href={url}
+        target="_blank"
+      >
         {url}
       </a>
       <p className={'pl-5 ' + TABLE_COLUMN_SIZE[3]}>{year}</p>
